Move instance map to vdom
